feat(frontend): getUser function pass into utils
